+++ src/chanjo2/endpoints/overview.py
@@ -1,5 +1,6 @@
+import datetime
 from os import path
-from typing import Dict, Optional
+from typing import Dict, Optional, Tuple
 
 from fastapi import APIRouter, Depends, Form, HTTPException, Request, status
 from fastapi.encoders import jsonable_encoder
@@ -11,7 +12,7 @@ from starlette.datastructures import FormData
 from typing_extensions import Annotated
 
 from chanjo2 import __version__
-from chanjo2.auth import get_current_user
+from chanjo2.auth import get_token
 from chanjo2.constants import DEFAULT_COVERAGE_LEVEL
 from chanjo2.dbutil import get_session
 from chanjo2.demo import DEMO_COVERAGE_QUERY_FORM, DEMO_GENE_OVERVIEW_QUERY_FORM
@@ -72,9 +73,10 @@ async def overview(
     hgnc_gene_symbols=Annotated[Optional[str], Form(None)],
     default_level=Annotated[Optional[int], Form(DEFAULT_COVERAGE_LEVEL)],
     db: Session = Depends(get_session),
-    user: dict = Depends(get_current_user),
+    token_data: Tuple[str, datetime.datetime] = Depends(get_token),
 ):
     """Return the genes overview page over a list of genes for a list of samples."""
+    validated_token, expires = token_data
     try:
         overview_query = ReportQuery.as_form(await request.form())
 
@@ -87,7 +89,7 @@ async def overview(
     overview_content: dict = get_report_data(
         query=overview_query, session=db, is_overview=True
     )
-    return templates.TemplateResponse(
+    response = templates.TemplateResponse(
         request=request,
         name="overview.html",
         context={
@@ -98,17 +100,29 @@ async def overview(
         },
     )
 
+    response.set_cookie(
+        key="id_token",
+        value=validated_token,
+        httponly=True,
+        secure=True,
+        samesite="Strict",
+        expires=expires,
+    )
+    return response
+
 
 @router.post("/gene_overview", response_class=HTMLResponse)
 async def gene_overview(
     request: Request,
     access_token=Annotated[Optional[str], Form(None)],
     db: Session = Depends(get_session),
-    user: dict = Depends(get_current_user),
+    token_data: Tuple[str, datetime.datetime] = Depends(get_token),
 ):
     """Returns coverage overview stats for a group of samples over genomic intervals of a single gene."""
+
     form_data: FormData = await request.form()
     form_dict: dict = jsonable_encoder(form_data)
+    validated_token, expires = token_data
 
     try:
         validated_form = GeneReportForm(**form_dict)
@@ -122,9 +136,19 @@ async def gene_overview(
         get_gene_overview_coverage_stats(form_data=validated_form, session=db)
     )
 
-    return templates.TemplateResponse(
+    response = templates.TemplateResponse(
         request=request, name="gene-overview.html", context=gene_overview_content
     )
+
+    response.set_cookie(
+        key="id_token",
+        value=validated_token,
+        httponly=True,
+        secure=True,
+        samesite="Strict",
+        expires=expires,
+    )
+    return response
 
 
 @router.get("/gene_overview/demo", response_class=HTMLResponse)
@@ -174,9 +198,10 @@ async def mane_overview(
     hgnc_gene_symbols=Annotated[Optional[str], Form(None)],
     default_level=Annotated[Optional[int], Form(DEFAULT_COVERAGE_LEVEL)],
     db: Session = Depends(get_session),
-    user: dict = Depends(get_current_user),
+    token_data: Tuple[str, datetime.datetime] = Depends(get_token),
 ):
     """Returns coverage overview stats for a group of samples over MANE transcripts of a list of genes."""
+    validated_token, expires = token_data
     try:
         overview_query = ReportQuery.as_form(await request.form())
 
@@ -186,8 +211,18 @@ async def mane_overview(
             detail=ve.json(),
         )
 
-    return templates.TemplateResponse(
+    response = templates.TemplateResponse(
         request=request,
         name="mane-overview.html",
         context=get_mane_overview_coverage_stats(query=overview_query, session=db),
     )
+
+    response.set_cookie(
+        key="id_token",
+        value=validated_token,
+        httponly=True,
+        secure=True,
+        samesite="Strict",
+        expires=expires,
+    )
+    return response